chore: Field type

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/group/v2/ApprovalDTO.kt
@@ -9,5 +9,5 @@ data class ApprovalDTO(
     @SerializedName("approval")
     val approval: Boolean?,
     @SerializedName("id")
-    val id: Int?
+    val id: Long?
 )
